gnu: gtk+-2: Support cross compilation.
gtk+-2 cannot be cross built, because it needs
an executable from itself. Specifically the one for updating icon cache.
To mitigate this requirement, this-package's bin output is added to
native-inputs when cross compiling.

This is only necessary for cross builds, for native builds, the same gtk+-2
is used. Care has been taken to not modify the package so that it produces the
same derivation as it did before.

* gnu/packages/gtk.scm (gtk+-2-for-build): Renamed from gtk+-2.
* gnu/packages/gtk.scm (gtk+-2): Inherit from gtk+-2, adding it to native-inputs.
[native-inputs]: For cross builds, add gtk+-2-for-build.
[arguments]{disallowed-references}: For cross builds, add gtk+-2-for-build.

Change-Id: I46c26764d52753bae67da055b96d77fd327052b6