gnu: make-nsis: Support cross compilation.
Here a hack has been added as there are two cross compilation toolchains
in the build process when cross compiling. The cross compile toolchain used
when using --target has been wrapped with the proper environment. The mingw
toolchain is using environment variables the same way it does so for native build.

* gnu/packages/installers.scm ((make-nsis machine target-arch nsis-target-type)): Update to 3.11.
[native-inputs]: Add bash-minimal.

Change-Id: Ieb64d5f5acbcf3fc9e17df5519344dda00cad53e