gnu: make-llvm: Disable tests for cross compilation.
The #:tests? argument has recently been added to make-llvm,
before the tests were enabled for native compilation and
disabled for cross compilation per the default arguments
in cmake-build and cmake-cross-build respectively.

But since this argument has been added, it overrides the
defaults and the test have been enabled on cross compilation,
mistakenly. The tests cannot run when cross compiling.

* gnu/packages/llvm.scm (make-llvm)[arguments]<#:tests?>:
Disable on cross compilation.

Change-Id: I5b3e33e6e9e1574eda5d1b74fe7484f157739dbb