build-system/python: Migrate to (json).
Since (json) is imported in Guix, we remove gradually the old (guix
build json) module.  This is not ideal for Python because the use of
the json module is very limited, and introduces the need to use
`with-extensions` at several places in the code.  At the same time,
keeping a complete json serialization for #:configure-flags keeps it
much more extensible than not doing that, and it's a good thing to
have a few examples in the codebase where `with-extensions` is used
properly.

* guix/build-system/pyproject.scm (%pyproject-build-system-modules):
Remove (guix build json).
(default-guile-json): New procedure.
(pyproject-build): Add guile-json extension, fix configure-flags
argument.

* guix/build/pyproject-build-system.scm (build): Refresh procedure
replacing (guix build json) procedures with (json) ones.