po: Do not handle escape sequences in PEG patterns.
\n sequences are handled and replaced by the 'interpret-newline-escape'
procedure.  Replacing cannot be done in PEG.

Before, PEG patterns had already dealt with the escaping of backslashes,
making it impossible for 'interpret-newline-escape' to distinguish \n
escape sequences from situations like \\n when the backslash itself is
escaped and thus cannot be part of an \n or \" escape sequence.

Escaping of backslashes must be done after \n and \", so it cannot be done
with PEG.

* guix/build/po.scm (str-chr): Do not escape characters.

Change-Id: Iae75dc146c55d21b725822780cfe951ef9ffae1e